requirements.txt
pandas==1.1.3
protobuf==3.20.*
google-cloud-bigquery==1.22.0